Add files via upload
Access the page for the TensorFlow Object Detection API. (https://github.com/opencv/opencv/wiki/TensorFlow-Object-Detection-API)

Locate and download the configuration file and the pretrained model weights for the Inception-SSD v2 model. To do this, right-click on the respective links and choose "Save As" to save the files to your computer.